requirements.txt
# ---- Core ----
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.9

# ---- Data & Imaging ----
opencv-python>=4.9.0
Pillow>=10.2.0
numpy>=1.26.0
scikit-image>=0.24.0
albumentations>=1.4.4

# ---- Metrics & Plots ----
scikit-learn>=1.4.0
matplotlib>=3.8.0
pandas>=2.2.0
seaborn>=0.13.0
joblib>=1.3.0          # for sklearn model persistence

# ---- Export / Runtime ----
onnx>=1.15.0
onnxruntime>=1.17.0
huggingface_hub>=0.22.2

# ---- Utils ----
tqdm>=4.66.0
pyyaml>=6.0.1
tensorboard>=2.16.0    # optional: training logs/visualization
torchsummary>=1.5.1    # optional: model summary debugging
